requirements.txt
pytest==8.4.2
playwright==1.58.0
pytest-xdist==3.8.0
allure-pytest==2.15.3
